+++ packages/commercial-lytenyte-pro/src/grid-box/item.tsx
@@ -1,7 +1,7 @@
-import { forwardRef, type JSX } from "react";
+import { forwardRef, type CSSProperties, type JSX } from "react";
 import { Item } from "../listbox/item";
 import type { GridBoxItem } from "./+types";
-import { DropWrap, useDraggable } from "@1771technologies/lytenyte-dragon";
+import { dragState, DropWrap, useDraggable } from "@1771technologies/lytenyte-dragon";
 import {
   useCombinedRefs,
   useSlot,
@@ -11,11 +11,13 @@ import { useGridBoxContext } from "./context";
 
 export interface GridBoxItemProps {
   readonly item: GridBoxItem;
-  readonly itemWrap?: SlotComponent;
+  readonly itemAs?: SlotComponent;
+  readonly itemClassName?: string;
+  readonly itemStyle?: CSSProperties;
 }
 
 export const BoxItem = forwardRef<HTMLDivElement, JSX.IntrinsicElements["div"] & GridBoxItemProps>(
-  function BoxItem({ item, itemWrap, ...props }, forwarded) {
+  function BoxItem({ item, itemAs, itemClassName, itemStyle, ...props }, forwarded) {
     const { accepted } = useGridBoxContext();
 
     const { dragProps } = useDraggable({
@@ -39,17 +41,55 @@ export const BoxItem = forwardRef<HTMLDivElement, JSX.IntrinsicElements["div"] &
               {...props}
               {...extraProps}
               onKeyDown={(ev) => {
-                if (ev.key === " ") item.onAction?.();
-                if (ev.key === "Backspace" || ev.key === "Delete") item.onDelete?.();
+                if (ev.key === " ") item.onAction?.(ev.currentTarget);
+                if (ev.key === "Backspace" || ev.key === "Delete")
+                  item.onDelete?.(ev.currentTarget);
               }}
-              onClick={() => item.onAction()}
+              onClick={(ev) => item.onAction(ev.currentTarget)}
+              className={itemClassName}
+              style={itemStyle}
             />
           ),
         },
       ],
-      slot: itemWrap ?? <div />,
+      slot: itemAs ?? <div />,
     });
 
-    return <DropWrap accepted={accepted} onDrop={item.onDrop} as={renderer} />;
+    return (
+      <DropWrap
+        onEnter={(e) => {
+          const data = dragState.active.get();
+
+          const thisSource = e.getAttribute("data-ln-source");
+          const dragSource = data?.getAttribute("data-ln-source");
+
+          if (!data) return;
+          if (thisSource !== dragSource) {
+            e.setAttribute("data-ln-is-after", "true");
+            return;
+          }
+
+          const thisIndex = Number.parseInt(e.getAttribute("data-ln-index")!);
+          const dragIndex = Number.parseInt(e.getAttribute("data-ln-index")!);
+
+          if (Number.isNaN(dragIndex) || Number.isNaN(thisIndex)) return;
+
+          if (thisIndex < dragIndex) {
+            e.setAttribute("data-ln-is-before", "true");
+          } else {
+            e.setAttribute("data-ln-is-after", "true");
+          }
+        }}
+        onLeave={(el) => {
+          el.removeAttribute("data-ln-is-before");
+          el.removeAttribute("data-ln-is-after");
+        }}
+        data-ln-source={item.source}
+        data-ln-index={item.index}
+        accepted={accepted}
+        onDrop={item.onDrop}
+        as={renderer}
+      />
+    );
   },
 );