+++ packages/lytenyte-core/src/state/use-lytenyte.ts
@@ -9,6 +9,7 @@ import type {
   PositionUnion,
   RowSelectionActivator,
   HeaderGroupCellLayout,
+  Column,
 } from "../+types.js";
 import { type Grid, type GridView, type UseLyteNyteProps } from "../+types.js";
 import { useRef } from "react";
@@ -151,6 +152,8 @@ export function makeLyteNyte<T>(p: UseLyteNyteProps<T>): Grid<T> {
   const virtualizeRows = atom(p.virtualizeRows ?? true);
   const virtualizeCols = atom(p.virtualizeCols ?? true);
 
+  const internal__rowGroupColumnState = atom<Record<string, Partial<Column<T>>>>({});
+
   const internal_rowSelectionPivot = atom<string | null>(null);
   const internal_rowSelectionLastWasDeselect = atom<boolean>(false);
   const rowSelectionPivot = atom((g) => g(internal_rowSelectionPivot));
@@ -225,11 +228,14 @@ export function makeLyteNyte<T>(p: UseLyteNyteProps<T>): Grid<T> {
    * is impacted by the column definitions, the group expansions, the row group display mode.
    */
   const columnView = atom((g) => {
+    const groupState = g(internal__rowGroupColumnState);
+
     const cols = columnAddRowGroup({
       columns: g(columns),
       rowGroupDisplayMode: g(rowGroupDisplayMode),
       rowGroupModel: g(rowGroupModel),
       rowGroupTemplate: g(rowGroupColumn),
+      rowGroupColumnState: groupState,
     });
 
     const colsWithMarker = columnHandleMarker({
@@ -539,7 +545,7 @@ export function makeLyteNyte<T>(p: UseLyteNyteProps<T>): Grid<T> {
     columnIndex: makeColumnIndex(grid),
     columnById: makeColumnById(grid),
     columnResize: makeColumnResize(grid),
-    columnUpdate: makeColumnUpdate(grid),
+    columnUpdate: makeColumnUpdate(grid as any),
     columnMove: makeColumnMove(grid),
     sortForColumn: makeSortForColumn(grid),
 
@@ -620,6 +626,7 @@ export function makeLyteNyte<T>(p: UseLyteNyteProps<T>): Grid<T> {
 
       draggingHeader: makeGridAtom(internal_draggingHeader, store),
 
+      rowGroupColumnState: makeGridAtom(internal__rowGroupColumnState, store),
       store: store,
     } satisfies InternalAtoms,
   });