+++ packages/commercial-lytenyte-license/CHANGELOG.md
@@ -1,5 +1,65 @@
 # @1771technologies/lng-license
 
+## 1.0.0-beta.14
+
+### Patch Changes
+
+- version
+
+## 1.0.0-beta.13
+
+### Patch Changes
+
+- version
+
+## 1.0.0-beta.12
+
+### Patch Changes
+
+- version
+
+## 1.0.0-beta.11
+
+### Patch Changes
+
+- version
+
+## 1.0.0-beta.10
+
+### Patch Changes
+
+- Version
+
+## 1.0.0-beta.9
+
+### Patch Changes
+
+- Version
+
+## 1.0.0-beta.8
+
+### Patch Changes
+
+- Version
+
+## 1.0.0-beta.7
+
+### Patch Changes
+
+- Fixes
+
+## 1.0.0-beta.6
+
+### Patch Changes
+
+- Version
+
+## 1.0.0-beta.5
+
+### Patch Changes
+
+- Version
+
 ## 1.0.0-beta.4
 
 ### Patch Changes