+++ packages/react-components/CHANGELOG.md
@@ -1,5 +1,165 @@
 # @1771technologies/lytenyte-react-components
 
+## 1.0.0-beta.14
+
+### Patch Changes
+
+- version
+- Updated dependencies
+  - @1771technologies/lytenyte-react-autocomplete@1.0.0-beta.14
+  - @1771technologies/lytenyte-scroll-lock@1.0.0-beta.14
+  - @1771technologies/lytenyte-focus-trap@1.0.0-beta.14
+  - @1771technologies/lytenyte-shared@1.0.0-beta.14
+  - @1771technologies/lytenyte-floating@1.0.0-beta.14
+  - @1771technologies/lytenyte-react-hooks@1.0.0-beta.14
+  - @1771technologies/lytenyte-focus@1.0.0-beta.14
+  - @1771technologies/lytenyte-dom-utils@1.0.0-beta.14
+  - @1771technologies/lytenyte-js-utils@1.0.0-beta.14
+
+## 1.0.0-beta.13
+
+### Patch Changes
+
+- version
+- Updated dependencies
+  - @1771technologies/lytenyte-react-autocomplete@1.0.0-beta.13
+  - @1771technologies/lytenyte-scroll-lock@1.0.0-beta.13
+  - @1771technologies/lytenyte-focus-trap@1.0.0-beta.13
+  - @1771technologies/lytenyte-shared@1.0.0-beta.13
+  - @1771technologies/lytenyte-floating@1.0.0-beta.13
+  - @1771technologies/lytenyte-react-hooks@1.0.0-beta.13
+  - @1771technologies/lytenyte-focus@1.0.0-beta.13
+  - @1771technologies/lytenyte-dom-utils@1.0.0-beta.13
+  - @1771technologies/lytenyte-js-utils@1.0.0-beta.13
+
+## 1.0.0-beta.12
+
+### Patch Changes
+
+- version
+- Updated dependencies
+  - @1771technologies/lytenyte-react-autocomplete@1.0.0-beta.12
+  - @1771technologies/lytenyte-scroll-lock@1.0.0-beta.12
+  - @1771technologies/lytenyte-focus-trap@1.0.0-beta.12
+  - @1771technologies/lytenyte-shared@1.0.0-beta.12
+  - @1771technologies/lytenyte-floating@1.0.0-beta.12
+  - @1771technologies/lytenyte-react-hooks@1.0.0-beta.12
+  - @1771technologies/lytenyte-focus@1.0.0-beta.12
+  - @1771technologies/lytenyte-dom-utils@1.0.0-beta.12
+  - @1771technologies/lytenyte-js-utils@1.0.0-beta.12
+
+## 1.0.0-beta.11
+
+### Patch Changes
+
+- version
+- Updated dependencies
+  - @1771technologies/lytenyte-react-autocomplete@1.0.0-beta.11
+  - @1771technologies/lytenyte-scroll-lock@1.0.0-beta.11
+  - @1771technologies/lytenyte-focus-trap@1.0.0-beta.11
+  - @1771technologies/lytenyte-shared@1.0.0-beta.11
+  - @1771technologies/lytenyte-floating@1.0.0-beta.11
+  - @1771technologies/lytenyte-react-hooks@1.0.0-beta.11
+  - @1771technologies/lytenyte-focus@1.0.0-beta.11
+  - @1771technologies/lytenyte-dom-utils@1.0.0-beta.11
+  - @1771technologies/lytenyte-js-utils@1.0.0-beta.11
+
+## 1.0.0-beta.10
+
+### Patch Changes
+
+- Version
+- Updated dependencies
+  - @1771technologies/lytenyte-react-autocomplete@1.0.0-beta.10
+  - @1771technologies/lytenyte-scroll-lock@1.0.0-beta.10
+  - @1771technologies/lytenyte-focus-trap@1.0.0-beta.10
+  - @1771technologies/lytenyte-shared@1.0.0-beta.10
+  - @1771technologies/lytenyte-floating@1.0.0-beta.10
+  - @1771technologies/lytenyte-react-hooks@1.0.0-beta.10
+  - @1771technologies/lytenyte-focus@1.0.0-beta.10
+  - @1771technologies/lytenyte-dom-utils@1.0.0-beta.10
+  - @1771technologies/lytenyte-js-utils@1.0.0-beta.10
+
+## 1.0.0-beta.9
+
+### Patch Changes
+
+- Version
+- Updated dependencies
+  - @1771technologies/lytenyte-react-autocomplete@1.0.0-beta.9
+  - @1771technologies/lytenyte-scroll-lock@1.0.0-beta.9
+  - @1771technologies/lytenyte-focus-trap@1.0.0-beta.9
+  - @1771technologies/lytenyte-shared@1.0.0-beta.9
+  - @1771technologies/lytenyte-floating@1.0.0-beta.9
+  - @1771technologies/lytenyte-react-hooks@1.0.0-beta.9
+  - @1771technologies/lytenyte-focus@1.0.0-beta.9
+  - @1771technologies/lytenyte-dom-utils@1.0.0-beta.9
+  - @1771technologies/lytenyte-js-utils@1.0.0-beta.9
+
+## 1.0.0-beta.8
+
+### Patch Changes
+
+- Version
+- Updated dependencies
+  - @1771technologies/lytenyte-react-autocomplete@1.0.0-beta.8
+  - @1771technologies/lytenyte-scroll-lock@1.0.0-beta.8
+  - @1771technologies/lytenyte-focus-trap@1.0.0-beta.8
+  - @1771technologies/lytenyte-shared@1.0.0-beta.8
+  - @1771technologies/lytenyte-floating@1.0.0-beta.8
+  - @1771technologies/lytenyte-react-hooks@1.0.0-beta.8
+  - @1771technologies/lytenyte-focus@1.0.0-beta.8
+  - @1771technologies/lytenyte-dom-utils@1.0.0-beta.8
+  - @1771technologies/lytenyte-js-utils@1.0.0-beta.8
+
+## 1.0.0-beta.7
+
+### Patch Changes
+
+- Fixes
+- Updated dependencies
+  - @1771technologies/lytenyte-react-autocomplete@1.0.0-beta.7
+  - @1771technologies/lytenyte-scroll-lock@1.0.0-beta.7
+  - @1771technologies/lytenyte-focus-trap@1.0.0-beta.7
+  - @1771technologies/lytenyte-shared@1.0.0-beta.7
+  - @1771technologies/lytenyte-floating@1.0.0-beta.7
+  - @1771technologies/lytenyte-react-hooks@1.0.0-beta.7
+  - @1771technologies/lytenyte-focus@1.0.0-beta.7
+  - @1771technologies/lytenyte-dom-utils@1.0.0-beta.7
+  - @1771technologies/lytenyte-js-utils@1.0.0-beta.7
+
+## 1.0.0-beta.6
+
+### Patch Changes
+
+- Version
+- Updated dependencies
+  - @1771technologies/lytenyte-react-autocomplete@1.0.0-beta.6
+  - @1771technologies/lytenyte-scroll-lock@1.0.0-beta.6
+  - @1771technologies/lytenyte-focus-trap@1.0.0-beta.6
+  - @1771technologies/lytenyte-shared@1.0.0-beta.6
+  - @1771technologies/lytenyte-floating@1.0.0-beta.6
+  - @1771technologies/lytenyte-react-hooks@1.0.0-beta.6
+  - @1771technologies/lytenyte-focus@1.0.0-beta.6
+  - @1771technologies/lytenyte-dom-utils@1.0.0-beta.6
+  - @1771technologies/lytenyte-js-utils@1.0.0-beta.6
+
+## 1.0.0-beta.5
+
+### Patch Changes
+
+- Version
+- Updated dependencies
+  - @1771technologies/lytenyte-react-autocomplete@1.0.0-beta.5
+  - @1771technologies/lytenyte-scroll-lock@1.0.0-beta.5
+  - @1771technologies/lytenyte-focus-trap@1.0.0-beta.5
+  - @1771technologies/lytenyte-shared@1.0.0-beta.5
+  - @1771technologies/lytenyte-floating@1.0.0-beta.5
+  - @1771technologies/lytenyte-react-hooks@1.0.0-beta.5
+  - @1771technologies/lytenyte-focus@1.0.0-beta.5
+  - @1771technologies/lytenyte-dom-utils@1.0.0-beta.5
+  - @1771technologies/lytenyte-js-utils@1.0.0-beta.5
+
 ## 1.0.0-beta.4
 
 ### Patch Changes